backend/requirements.txt
fastapi
uvicorn
openai
httpx
python-dotenv
python-multipart
pypdf
pydantic
